--- a/src/liquid/context.cr
+++ b/src/liquid/context.cr
@@ -82,59 +82,64 @@ module Liquid
         next unless k
 
         # ignore array index/hash key if present, first handle methods/properties
-        if k =~ /^(.*?)(?:\[.*?\])*$/
-          name = $1
-          if ret
-            # handle a selection of convenience methods
-            # wrap values in Any to make sure the return type is always the same
-            case name
-            when "present"
-              if (array = ret.as_a?)
-                ret = Any.new(array.size > 0)
-              elsif (hash = ret.as_h?)
-                ret = Any.new(hash.keys.size > 0)
-              else
-                ret = Any.new(ret.raw.to_s.size > 0)
-              end
-            when "blank"
-              if (array = ret.as_a?)
-                ret = Any.new(array.size == 0)
-              elsif (hash = ret.as_h?)
-                ret = Any.new(hash.keys.size == 0)
-              else
-                ret = Any.new(ret.raw.to_s.size == 0)
-              end
-            when "size"
-              if (array = ret.as_a?)
-                ret = Any.new(array.size)
-              elsif (hash = ret.as_h?)
-                ret = Any.new(hash.keys.size)
-              elsif (str = ret.as_s?)
-                ret = Any.new(str.size)
-              else
-                return parse_error(key, strict, "Parse error: Tried to call #size on something other than a String, Array or Hash (#{key} -> #{ret.inspect})")
-              end
-            else
-              # if not a method, then it's a property (implemented as JSON hash member)
-              if (hash = ret.as_h?)
-                return key_missing(key, strict) unless hash.has_key?(k)
+        bracket_index = k.index('[')
+        name =
+          if bracket_index && k.ends_with?(']')
+            k[0...bracket_index]
+          else
+            k
+          end
 
-                ret = hash[k]
-              else
-                return parse_error(key, strict, "Parse error: Tried to access property of a non-hash object (#{key} -> #{ret.inspect})")
-              end
+        if ret
+          # handle a selection of convenience methods
+          # wrap values in Any to make sure the return type is always the same
+          case name
+          when "present"
+            if (array = ret.as_a?)
+              ret = Any.new(array.size > 0)
+            elsif (hash = ret.as_h?)
+              ret = Any.new(hash.keys.size > 0)
+            else
+              ret = Any.new(ret.raw.to_s.size > 0)
+            end
+          when "blank"
+            if (array = ret.as_a?)
+              ret = Any.new(array.size == 0)
+            elsif (hash = ret.as_h?)
+              ret = Any.new(hash.keys.size == 0)
+            else
+              ret = Any.new(ret.raw.to_s.size == 0)
+            end
+          when "size"
+            if (array = ret.as_a?)
+              ret = Any.new(array.size)
+            elsif (hash = ret.as_h?)
+              ret = Any.new(hash.keys.size)
+            elsif (str = ret.as_s?)
+              ret = Any.new(str.size)
+            else
+              return parse_error(key, strict, "Parse error: Tried to call #size on something other than a String, Array or Hash (#{key} -> #{ret.inspect})")
             end
           else
-            # first time through ret = nil, name should correspond to a top level key in @inner
-            if @inner.has_key?(name)
-              ret = @inner[name]
+            # if not a method, then it's a property (implemented as JSON hash member)
+            if (hash = ret.as_h?)
+              return key_missing(key, strict) unless hash.has_key?(k)
+
+              ret = hash[k]
             else
-              if strict
-                return key_missing(key, strict)
-              else
-                # keep going, in case we're ultimately just doing a #blank check
-                ret = Any.new(nil)
-              end
+              return parse_error(key, strict, "Parse error: Tried to access property of a non-hash object (#{key} -> #{ret.inspect})")
+            end
+          end
+        else
+          # first time through ret = nil, name should correspond to a top level key in @inner
+          if @inner.has_key?(name)
+            ret = @inner[name]
+          else
+            if strict
+              return key_missing(key, strict)
+            else
+              # keep going, in case we're ultimately just doing a #blank check
+              ret = Any.new(nil)
             end
           end
         end