@asterite 2022-11-08
There's room for more optimizations, but this is the simplest thing that can be done (other optimizations will take me a bit more time.)