@crimson-knight 2022-11-08
@asterite microseconds is definitely what I like to see with a crystal shards processing time!