@asterite 2022-11-08
I just pushed another small commit. Now the times are like this:

```
parse & render  25.55k ( 39.15µs) (± 1.30%)  49.7kB/op   1.79× slower
        render  45.77k ( 21.85µs) (± 0.98%)  22.2kB/op        fastest
```

So now it's 40~46 times faster than before.

There are more regexes that could be replaced with faster logic, but I think for now this is good enough.